% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitstream-core.R
\name{assert_bs}
\alias{assert_bs}
\title{Test if an object is a bitstream object and fail if it is not}
\usage{
assert_bs(x)
}
\arguments{
\item{x}{object to test}
}
\value{
None
}
\description{
Test if an object is a bitstream object and fail if it is not
}
\examples{
raw_vec <- as.raw(1:3)
bs  <- bs_open(raw_vec, 'r')
assert_bs(bs)
bs_close(bs)

}
