% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitstream-core.R
\name{bs_advance}
\alias{bs_advance}
\title{Advance bitstream}
\usage{
bs_advance(bs, n)
}
\arguments{
\item{bs}{Bistream connection object created with \code{bs_open()}}

\item{n}{number of bits to advance}
}
\value{
\code{Bitstream} connection returned invisibly
}
\description{
Advance bitstream
}
\examples{
raw_vec <- as.raw(1:3)
bs  <- bs_open(raw_vec, 'r')
bs_is_aligned(bs)
bs_advance(bs, 4)
bs_is_aligned(bs)
bs_read_bit(bs, 8)
bs_is_aligned(bs)
bs_close(bs)
}
