% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitstream-core.R
\name{bs_flush}
\alias{bs_flush}
\title{Flush bits in the buffer}
\usage{
bs_flush(bs)
}
\arguments{
\item{bs}{Bistream connection object created with \code{bs_open()}}
}
\value{
\code{Bitstream} connection returned invisibly
}
\description{
This is called internally to flush bitstream buffers to the underlying
R connection.
}
\examples{
bs  <- bs_open(raw(), 'w')
bs_write_bit(bs, c(TRUE, FALSE, TRUE))
bs_align(bs, nbits = 8)
bs_flush(bs)
output <- bs_close(bs)
output
}
