% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitstream-core.R
\name{bs_read_bit}
\alias{bs_read_bit}
\title{Read bits from a bitstream}
\usage{
bs_read_bit(bs, n)
}
\arguments{
\item{bs}{Bistream connection object created with \code{bs_open()}}

\item{n}{number of bits to read}
}
\value{
logical vector of bit values
}
\description{
Read bits from a bitstream
}
\examples{
raw_vec <- as.raw(1:3)
bs  <- bs_open(raw_vec, 'r')
bs_read_bit(bs, 4)
bs_is_aligned(bs)
bs_close(bs)
}
