% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitstream-core.R
\name{bs_write_bit}
\alias{bs_write_bit}
\title{Write unaligned bits to a bitstream}
\usage{
bs_write_bit(bs, x)
}
\arguments{
\item{bs}{Bistream connection object created with \code{bs_open()}}

\item{x}{Logical vector of bit values}
}
\value{
\code{Bitstream} connection returned invisibly
}
\description{
Write unaligned bits to a bitstream
}
\examples{
bs  <- bs_open(raw(), 'w')
bs_write_bit(bs, c(TRUE, FALSE, TRUE))
bs_align(bs, nbits = 8)
bs_flush(bs) 
output <- bs_close(bs)
output
}
