% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitstream-core.R
\name{is_bs}
\alias{is_bs}
\title{Test if an object is a bitstream object}
\usage{
is_bs(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical. TRUE if object is a bitstream object
}
\description{
Test if an object is a bitstream object
}
\examples{
# Negative case
is_bs(NULL)

# Positive case
raw_vec <- as.raw(1:3)
bs  <- bs_open(raw_vec, 'r')
is_bs(bs)
bs_close(bs)
}
