% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_bits}
\alias{pad_bits}
\title{Pad a logical vector to the given size}
\usage{
pad_bits(bits, nbits = 8L, side = "left", value = FALSE)
}
\arguments{
\item{bits}{logical vector}

\item{nbits}{Pad width to a multiple of this number of bits}

\item{side}{'left' or 'right'.  Only the lowercase version of the
first letter is used to determine the side. all of these are valid 
options: 'L', 'R', 'left', 'Right'}

\item{value}{The value to use for padding. single logical value. 
Default: FALSE}
}
\value{
Logical vector with the specified padding
}
\description{
Pad a logical vector to the given size
}
\examples{
pad_bits(c(TRUE, TRUE), nbits = 8, side = 'left')
pad_bits(c(TRUE, TRUE), nbits = 8, side = 'R')

}
